Move cms/* to ddevapp, build out apptypes.go to switch functions for apptypes
