--- pkg/cms/config/drupal.go
@@ -1,125 +1,0 @@
-package config
-
-import (
-	"os"
-	"text/template"
-
-	"path/filepath"
-
-	"github.com/Masterminds/sprig"
-	"github.com/drud/ddev/pkg/cms/model"
-	"github.com/drud/ddev/pkg/util"
-)
-
-const (
-	drupalTemplate = `<?php
-{{ $config := . }}
-/**
- {{ $config.Signature }}: Automatically generated Drupal settings.php file.
- ddev manages this file and may delete or overwrite the file unless this comment is removed.
- */
-
-$databases['default']['default'] = array(
-  'database' => "{{ $config.DatabaseName }}",
-  'username' => "{{ $config.DatabaseUsername }}",
-  'password' => "{{ $config.DatabasePassword }}",
-  'host' => "{{ $config.DatabaseHost }}",
-  'driver' => "{{ $config.DatabaseDriver }}",
-  'port' => {{ $config.DatabasePort }},
-  'prefix' => "{{ $config.DatabasePrefix }}",
-);
-
-ini_set('session.gc_probability', 1);
-ini_set('session.gc_divisor', 100);
-ini_set('session.gc_maxlifetime', 200000);
-ini_set('session.cookie_lifetime', 2000000);
-
-{{ if $config.IsDrupal8 }}
-
-$settings['hash_salt'] = '{{ $config.HashSalt }}';
-
-$settings['file_scan_ignore_directories'] = [
-  'node_modules',
-  'bower_components',
-];
-
-
-{{ else }}
-
-$drupal_hash_salt = '{{ $config.HashSalt }}';
-{{ end }}
-
-
-// This is super ugly but it determines whether or not drush should include a custom settings file which allows
-// it to work both within a docker container and natively on the host system.
-if (!empty($_SERVER["argv"]) && strpos($_SERVER["argv"][0], "drush") && empty($_ENV['DEPLOY_NAME'])) {
-  include __DIR__ . '../../../drush.settings.php';
-}
-`
-)
-
-const drushTemplate = `<?php
-{{ $config := . }}
-$databases['default']['default'] = array(
-  'database' => "db",
-  'username' => "db",
-  'password' => "db",
-  'host' => "127.0.0.1",
-  'driver' => "mysql",
-  'port' => {{ $config.DatabasePort }},
-  'prefix' => "",
-);
-`
-
-// WriteDrupalConfig dynamically produces valid settings.php file by combining a configuration
-// object with a data-driven template.
-func WriteDrupalConfig(drupalConfig *model.DrupalConfig, filePath string) error {
-	tmpl, err := template.New("drupalConfig").Funcs(sprig.TxtFuncMap()).Parse(drupalTemplate)
-	if err != nil {
-		return err
-	}
-
-	// Ensure target directory is writable.
-	dir := filepath.Dir(filePath)
-	err = os.Chmod(dir, 0755)
-	if err != nil {
-		return err
-	}
-
-	file, err := os.Create(filePath)
-	if err != nil {
-		return err
-	}
-	err = tmpl.Execute(file, drupalConfig)
-	if err != nil {
-		return err
-	}
-	util.CheckClose(file)
-	return nil
-}
-
-// WriteDrushConfig writes out a drush config based on passed-in values.
-func WriteDrushConfig(drushConfig *model.DrushConfig, filePath string) error {
-	tmpl, err := template.New("drushConfig").Funcs(sprig.TxtFuncMap()).Parse(drushTemplate)
-	if err != nil {
-		return err
-	}
-
-	// Ensure target directory is writable.
-	dir := filepath.Dir(filePath)
-	err = os.Chmod(dir, 0755)
-	if err != nil {
-		return err
-	}
-
-	file, err := os.Create(filePath)
-	if err != nil {
-		return err
-	}
-	err = tmpl.Execute(file, drushConfig)
-	if err != nil {
-		return err
-	}
-	util.CheckClose(file)
-	return nil
-}
