--- pkg/cms/model/drupal.go
@@ -1,54 +1,0 @@
-package model
-
-import (
-	"github.com/drud/ddev/pkg/appports"
-	"github.com/drud/ddev/pkg/util"
-)
-
-// DrupalConfig encapsulates all the configurations for a Drupal site.
-type DrupalConfig struct {
-	DeployName       string
-	DeployURL        string
-	DatabaseName     string
-	DatabaseUsername string
-	DatabasePassword string
-	DatabaseHost     string
-	DatabaseDriver   string
-	DatabasePort     string
-	DatabasePrefix   string
-	HashSalt         string
-	IsDrupal8        bool
-	Signature        string
-}
-
-// NewDrupalConfig produces a DrupalConfig object with default.
-func NewDrupalConfig() *DrupalConfig {
-	return &DrupalConfig{
-		DatabaseName:     "db",
-		DatabaseUsername: "db",
-		DatabasePassword: "db",
-		DatabaseHost:     "db",
-		DatabaseDriver:   "mysql",
-		DatabasePort:     appports.GetPort("db"),
-		DatabasePrefix:   "",
-		IsDrupal8:        false,
-		HashSalt:         util.RandString(64),
-		Signature:        DdevSettingsFileSignature,
-	}
-}
-
-// DrushConfig encapsulates configuration for a drush settings file.
-type DrushConfig struct {
-	DatabasePort string
-	DatabaseHost string
-	IsDrupal8    bool
-}
-
-// NewDrushConfig produces a DrushConfig object with default.
-func NewDrushConfig() *DrushConfig {
-	return &DrushConfig{
-		DatabaseHost: "127.0.0.1",
-		DatabasePort: appports.GetPort("db"),
-		IsDrupal8:    false,
-	}
-}
