--- pkg/cms/model/wordpress.go
@@ -1,49 +1,0 @@
-package model
-
-import (
-	"github.com/drud/ddev/pkg/util"
-)
-
-// WordpressConfig encapsulates all the configurations for a WordPress site.
-type WordpressConfig struct {
-	WPGeneric        bool
-	DeployName       string
-	DeployURL        string
-	DatabaseName     string
-	DatabaseUsername string
-	DatabasePassword string
-	DatabaseHost     string
-	AuthKey          string
-	SecureAuthKey    string
-	LoggedInKey      string
-	NonceKey         string
-	AuthSalt         string
-	SecureAuthSalt   string
-	LoggedInSalt     string
-	NonceSalt        string
-	Docroot          string
-	TablePrefix      string
-	Signature        string
-}
-
-// NewWordpressConfig produces a WordpressConfig object with defaults.
-func NewWordpressConfig() *WordpressConfig {
-	return &WordpressConfig{
-		WPGeneric:        false,
-		DatabaseName:     "db",
-		DatabaseUsername: "db",
-		DatabasePassword: "db",
-		DatabaseHost:     "db",
-		Docroot:          "/var/www/html/docroot",
-		TablePrefix:      "wp_",
-		AuthKey:          util.RandString(64),
-		AuthSalt:         util.RandString(64),
-		LoggedInKey:      util.RandString(64),
-		LoggedInSalt:     util.RandString(64),
-		NonceKey:         util.RandString(64),
-		NonceSalt:        util.RandString(64),
-		SecureAuthKey:    util.RandString(64),
-		SecureAuthSalt:   util.RandString(64),
-		Signature:        DdevSettingsFileSignature,
-	}
-}
